Check for Changes Between Two Query Reports

This function compares an old report of queries with a new one. It allows you to identify which queries are new, which have been modified, and which remain unchanged.

Usage

check_queries(old, new, report_title = NULL)

Arguments

old

Previous version of the queries report.

new

New version of the queries report. This object is used to determine the status of each query.

report_title

Character string specifying the title of the report.

Value

A list consisting of a dataframe containing each individual query from both reports and a column showing the status of the queries (new, solved, miscorrected or pending) compared to the previous query report. In addition to this dataframe, there is also a summary of the total number of queries per category.

Examples

# Example of a query
data_old <- rd_query(covican,
                     variables = "copd",
                     expression = "is.na(x)",
                     event = "baseline_visit_arm_1")
data_new <- rbind(data_old$queries[1:5,], c("100-20",rep("abc",8)))

# Control of queries
check <- check_queries(old = data_old$queries,
                       new = data_new)