Identification of queries

This function allows you to identify queries by the use a specific expression. It can be used to identify missing values or to identify values that are outside the lower and upper limits of a variable.

Usage

rd_query(
  ...,
  data = NULL,
  dic = NULL,
  variables = NA,
  expression = NA,
  negate = FALSE,
  event = NA,
  filter = NA,
  addTo = NA,
  variables_names = NA,
  query_name = NA,
  instrument = NA,
  report_title = NA,
  report_zeros = FALSE,
  by_dag = FALSE,
  link = list()
)

Arguments

...

List containing the data, the dictionary and the event (if required). It may be the output of the `redcap_data` function.

data

Data frame containing the data read from REDCap. If the list is given, this argument is not required.

dic

Data frame containing the dictionary read from REDCap. If the list is given, this argument is not required.

variables

Character vector of the names of the database variables to be checked.

expression

Character vector of expressions to be applied to the selected variables.

negate

Logical value which indicates whether the defined expression should be negated. Defaults to `FALSE`.

event

The name of the REDCap event to be analyzed. If there are events in your REDCap project, you should use this argument in order to name the event to which the defined variables belong.

filter

A filter to be applied to the dataset. For example, the branching logic of a determined variable can be applied using this argument.

addTo

Data frame corresponding to a previous query data frame to which you can add the new query data frame. By default, the function always generates a new data frame without taking into account previous reports.

variables_names

Character vector containing the description of each selected variable. By default, the function automatically takes the description of each variable from the dictionary of the REDCap project.

query_name

Description of the query. It can be defined as the same one for all the variables, or you can define a different one for each variable. By default, the function defines it as `The value is [value] and it should not be [expression]'`.

instrument

REDCap's instrument to which the variables belong. It can be defined as the same one for all the variables, or you can define a different one for each variable. By default, the function automatically selects the corresponding instrument of each variable from the dictionary of the REDCap project.

report_title

Character string specifying the title of the report.

report_zeros

Logical. If `TRUE`, the function returns a report containing variables with zero queries.

by_dag

Logical. If `TRUE`, both elements of the output are grouped by the data access groups (DAGs) of the REDCap project.

link

List containing project information used to create a web link to each query.

Value

A list with a data frame of 9 columns (10 columns, if the link argument is specified) meant to help the user identify each query and a table with the total number of queries per variable.

Examples

# Missing values
example <- rd_query(covican,
                    variables = c("copd", "age"),
                    expression = c("%in%NA", "%in%NA"),
                    event = "baseline_visit_arm_1")
example
#> $queries
#>    Identifier         DAG          Event    Instrument Field Repetition
#> 1      100-58 Hospital 11 Baseline visit Comorbidities  copd          -
#> 2     102-113 Hospital 24 Baseline visit  Demographics   age          -
#> 3      105-11  Hospital 5 Baseline visit Comorbidities  copd          -
#> 4      105-11  Hospital 5 Baseline visit  Demographics   age          -
#> 5      105-56  Hospital 5 Baseline visit Comorbidities  copd          -
#> 6      105-56  Hospital 5 Baseline visit  Demographics   age          -
#> 7      112-15 Hospital 10 Baseline visit Comorbidities  copd          -
#> 8      117-11  Hospital 2 Baseline visit Comorbidities  copd          -
#> 9      117-11  Hospital 2 Baseline visit  Demographics   age          -
#> 10     117-22  Hospital 2 Baseline visit Comorbidities  copd          -
#> 11     117-22  Hospital 2 Baseline visit  Demographics   age          -
#>                              Description
#> 1  Chronic obstructive pulmonary disease
#> 2                                    Age
#> 3  Chronic obstructive pulmonary disease
#> 4                                    Age
#> 5  Chronic obstructive pulmonary disease
#> 6                                    Age
#> 7  Chronic obstructive pulmonary disease
#> 8  Chronic obstructive pulmonary disease
#> 9                                    Age
#> 10 Chronic obstructive pulmonary disease
#> 11                                   Age
#>                                           Query      Code
#> 1  The value is NA and it should not be missing  100-58-1
#> 2  The value is NA and it should not be missing 102-113-1
#> 3  The value is NA and it should not be missing  105-11-1
#> 4  The value is NA and it should not be missing  105-11-2
#> 5  The value is NA and it should not be missing  105-56-1
#> 6  The value is NA and it should not be missing  105-56-2
#> 7  The value is NA and it should not be missing  112-15-1
#> 8  The value is NA and it should not be missing  117-11-1
#> 9  The value is NA and it should not be missing  117-11-2
#> 10 The value is NA and it should not be missing  117-22-1
#> 11 The value is NA and it should not be missing  117-22-2
#> 
#> $results
#> <table class="table table-striped table-condensed" style="width: auto !important; margin-left: auto; margin-right: auto;">
#> <caption>Report of queries</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Variables </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Description </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Event </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Query </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Total </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:center;"> copd </td>
#>    <td style="text-align:center;"> Chronic obstructive pulmonary disease </td>
#>    <td style="text-align:center;"> Baseline visit </td>
#>    <td style="text-align:center;"> The value should not be missing </td>
#>    <td style="text-align:center;"> 6 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:center;"> age </td>
#>    <td style="text-align:center;"> Age </td>
#>    <td style="text-align:center;"> Baseline visit </td>
#>    <td style="text-align:center;"> The value should not be missing </td>
#>    <td style="text-align:center;"> 5 </td>
#>   </tr>
#> </tbody>
#> </table>

# Expression
example <- rd_query(covican,
                    variables="age",
                    expression=">20",
                    event="baseline_visit_arm_1")
example
#> $queries
#>     Identifier         DAG          Event   Instrument Field Repetition
#> 1        100-6 Hospital 11 Baseline visit Demographics   age          -
#> 2       100-13 Hospital 11 Baseline visit Demographics   age          -
#> 3       100-16 Hospital 11 Baseline visit Demographics   age          -
#> 4       100-31 Hospital 11 Baseline visit Demographics   age          -
#> 5       100-34 Hospital 11 Baseline visit Demographics   age          -
#> 6       100-36 Hospital 11 Baseline visit Demographics   age          -
#> 7       100-52 Hospital 11 Baseline visit Demographics   age          -
#> 8       100-58 Hospital 11 Baseline visit Demographics   age          -
#> 9       100-65 Hospital 11 Baseline visit Demographics   age          -
#> 10      100-66 Hospital 11 Baseline visit Demographics   age          -
#> 11      100-72 Hospital 11 Baseline visit Demographics   age          -
#> 12      100-82 Hospital 11 Baseline visit Demographics   age          -
#> 13       101-8  Hospital 1 Baseline visit Demographics   age          -
#> 14      101-13  Hospital 1 Baseline visit Demographics   age          -
#> 15      101-16  Hospital 1 Baseline visit Demographics   age          -
#> 16      101-19  Hospital 1 Baseline visit Demographics   age          -
#> 17      101-22  Hospital 1 Baseline visit Demographics   age          -
#> 18      101-36  Hospital 1 Baseline visit Demographics   age          -
#> 19      101-45  Hospital 1 Baseline visit Demographics   age          -
#> 20      101-48  Hospital 1 Baseline visit Demographics   age          -
#> 21      101-59  Hospital 1 Baseline visit Demographics   age          -
#> 22      101-70  Hospital 1 Baseline visit Demographics   age          -
#> 23      101-74  Hospital 1 Baseline visit Demographics   age          -
#> 24      101-75  Hospital 1 Baseline visit Demographics   age          -
#> 25      101-78  Hospital 1 Baseline visit Demographics   age          -
#> 26      101-83  Hospital 1 Baseline visit Demographics   age          -
#> 27      101-85  Hospital 1 Baseline visit Demographics   age          -
#> 28      101-86  Hospital 1 Baseline visit Demographics   age          -
#> 29      101-88  Hospital 1 Baseline visit Demographics   age          -
#> 30      101-93  Hospital 1 Baseline visit Demographics   age          -
#> 31      101-96  Hospital 1 Baseline visit Demographics   age          -
#> 32      101-98  Hospital 1 Baseline visit Demographics   age          -
#> 33       102-6 Hospital 24 Baseline visit Demographics   age          -
#> 34      102-10 Hospital 24 Baseline visit Demographics   age          -
#> 35      102-13 Hospital 24 Baseline visit Demographics   age          -
#> 36      102-15 Hospital 24 Baseline visit Demographics   age          -
#> 37      102-23 Hospital 24 Baseline visit Demographics   age          -
#> 38      102-27 Hospital 24 Baseline visit Demographics   age          -
#> 39      102-30 Hospital 24 Baseline visit Demographics   age          -
#> 40      102-39 Hospital 24 Baseline visit Demographics   age          -
#> 41      102-48 Hospital 24 Baseline visit Demographics   age          -
#> 42      102-60 Hospital 24 Baseline visit Demographics   age          -
#> 43      102-64 Hospital 24 Baseline visit Demographics   age          -
#> 44      102-73 Hospital 24 Baseline visit Demographics   age          -
#> 45      102-74 Hospital 24 Baseline visit Demographics   age          -
#> 46      102-75 Hospital 24 Baseline visit Demographics   age          -
#> 47      102-83 Hospital 24 Baseline visit Demographics   age          -
#> 48      102-84 Hospital 24 Baseline visit Demographics   age          -
#> 49     102-103 Hospital 24 Baseline visit Demographics   age          -
#> 50     102-104 Hospital 24 Baseline visit Demographics   age          -
#> 51     102-127 Hospital 24 Baseline visit Demographics   age          -
#> 52     102-134 Hospital 24 Baseline visit Demographics   age          -
#> 53       103-6 Hospital 20 Baseline visit Demographics   age          -
#> 54       103-9 Hospital 20 Baseline visit Demographics   age          -
#> 55      103-10 Hospital 20 Baseline visit Demographics   age          -
#> 56      103-14 Hospital 20 Baseline visit Demographics   age          -
#> 57      103-16 Hospital 20 Baseline visit Demographics   age          -
#> 58      103-17 Hospital 20 Baseline visit Demographics   age          -
#> 59      103-19 Hospital 20 Baseline visit Demographics   age          -
#> 60       104-8 Hospital 25 Baseline visit Demographics   age          -
#> 61      104-12 Hospital 25 Baseline visit Demographics   age          -
#> 62      104-13 Hospital 25 Baseline visit Demographics   age          -
#> 63       105-8  Hospital 5 Baseline visit Demographics   age          -
#> 64       105-9  Hospital 5 Baseline visit Demographics   age          -
#> 65      105-13  Hospital 5 Baseline visit Demographics   age          -
#> 66      105-19  Hospital 5 Baseline visit Demographics   age          -
#> 67      105-21  Hospital 5 Baseline visit Demographics   age          -
#> 68      105-22  Hospital 5 Baseline visit Demographics   age          -
#> 69      105-30  Hospital 5 Baseline visit Demographics   age          -
#> 70      105-46  Hospital 5 Baseline visit Demographics   age          -
#> 71      105-52  Hospital 5 Baseline visit Demographics   age          -
#> 72      105-58  Hospital 5 Baseline visit Demographics   age          -
#> 73      105-78  Hospital 5 Baseline visit Demographics   age          -
#> 74      105-85  Hospital 5 Baseline visit Demographics   age          -
#> 75       106-8 Hospital 19 Baseline visit Demographics   age          -
#> 76       106-9 Hospital 19 Baseline visit Demographics   age          -
#> 77      106-15 Hospital 19 Baseline visit Demographics   age          -
#> 78      106-18 Hospital 19 Baseline visit Demographics   age          -
#> 79      106-22 Hospital 19 Baseline visit Demographics   age          -
#> 80      106-27 Hospital 19 Baseline visit Demographics   age          -
#> 81      106-29 Hospital 19 Baseline visit Demographics   age          -
#> 82      106-32 Hospital 19 Baseline visit Demographics   age          -
#> 83      106-33 Hospital 19 Baseline visit Demographics   age          -
#> 84      107-11  Hospital 6 Baseline visit Demographics   age          -
#> 85       108-8 Hospital 18 Baseline visit Demographics   age          -
#> 86      108-13 Hospital 18 Baseline visit Demographics   age          -
#> 87      108-16 Hospital 18 Baseline visit Demographics   age          -
#> 88      108-17 Hospital 18 Baseline visit Demographics   age          -
#> 89      108-19 Hospital 18 Baseline visit Demographics   age          -
#> 90      108-23 Hospital 18 Baseline visit Demographics   age          -
#> 91      108-28 Hospital 18 Baseline visit Demographics   age          -
#> 92      108-29 Hospital 18 Baseline visit Demographics   age          -
#> 93      108-30 Hospital 18 Baseline visit Demographics   age          -
#> 94      108-35 Hospital 18 Baseline visit Demographics   age          -
#> 95      108-44 Hospital 18 Baseline visit Demographics   age          -
#> 96      108-45 Hospital 18 Baseline visit Demographics   age          -
#> 97      108-56 Hospital 18 Baseline visit Demographics   age          -
#> 98      108-63 Hospital 18 Baseline visit Demographics   age          -
#> 99       109-8 Hospital 21 Baseline visit Demographics   age          -
#> 100     109-10 Hospital 21 Baseline visit Demographics   age          -
#> 101     109-11 Hospital 21 Baseline visit Demographics   age          -
#> 102     109-19 Hospital 21 Baseline visit Demographics   age          -
#> 103     109-21 Hospital 21 Baseline visit Demographics   age          -
#> 104     109-22 Hospital 21 Baseline visit Demographics   age          -
#> 105     109-32 Hospital 21 Baseline visit Demographics   age          -
#> 106     109-33 Hospital 21 Baseline visit Demographics   age          -
#> 107     109-34 Hospital 21 Baseline visit Demographics   age          -
#> 108     109-35 Hospital 21 Baseline visit Demographics   age          -
#> 109     109-37 Hospital 21 Baseline visit Demographics   age          -
#> 110     109-38 Hospital 21 Baseline visit Demographics   age          -
#> 111      110-6 Hospital 26 Baseline visit Demographics   age          -
#> 112     110-19 Hospital 26 Baseline visit Demographics   age          -
#> 113      111-7 Hospital 14 Baseline visit Demographics   age          -
#> 114      111-8 Hospital 14 Baseline visit Demographics   age          -
#> 115     111-10 Hospital 14 Baseline visit Demographics   age          -
#> 116      112-6 Hospital 10 Baseline visit Demographics   age          -
#> 117      112-7 Hospital 10 Baseline visit Demographics   age          -
#> 118     112-10 Hospital 10 Baseline visit Demographics   age          -
#> 119     112-11 Hospital 10 Baseline visit Demographics   age          -
#> 120     112-12 Hospital 10 Baseline visit Demographics   age          -
#> 121     112-14 Hospital 10 Baseline visit Demographics   age          -
#> 122     112-15 Hospital 10 Baseline visit Demographics   age          -
#> 123     112-18 Hospital 10 Baseline visit Demographics   age          -
#> 124     112-19 Hospital 10 Baseline visit Demographics   age          -
#> 125     112-20 Hospital 10 Baseline visit Demographics   age          -
#> 126     112-21 Hospital 10 Baseline visit Demographics   age          -
#> 127      113-9  Hospital 8 Baseline visit Demographics   age          -
#> 128     113-10  Hospital 8 Baseline visit Demographics   age          -
#> 129     113-11  Hospital 8 Baseline visit Demographics   age          -
#> 130     113-12  Hospital 8 Baseline visit Demographics   age          -
#> 131     113-13  Hospital 8 Baseline visit Demographics   age          -
#> 132     113-16  Hospital 8 Baseline visit Demographics   age          -
#> 133     113-21  Hospital 8 Baseline visit Demographics   age          -
#> 134     113-23  Hospital 8 Baseline visit Demographics   age          -
#> 135     113-24  Hospital 8 Baseline visit Demographics   age          -
#> 136     113-27  Hospital 8 Baseline visit Demographics   age          -
#> 137     113-31  Hospital 8 Baseline visit Demographics   age          -
#> 138     113-32  Hospital 8 Baseline visit Demographics   age          -
#> 139     113-36  Hospital 8 Baseline visit Demographics   age          -
#> 140     113-37  Hospital 8 Baseline visit Demographics   age          -
#> 141     113-38  Hospital 8 Baseline visit Demographics   age          -
#> 142     113-40  Hospital 8 Baseline visit Demographics   age          -
#> 143     113-41  Hospital 8 Baseline visit Demographics   age          -
#> 144     113-44  Hospital 8 Baseline visit Demographics   age          -
#> 145     114-13 Hospital 13 Baseline visit Demographics   age          -
#> 146     114-17 Hospital 13 Baseline visit Demographics   age          -
#> 147     114-19 Hospital 13 Baseline visit Demographics   age          -
#> 148      115-8 Hospital 15 Baseline visit Demographics   age          -
#> 149     115-13 Hospital 15 Baseline visit Demographics   age          -
#> 150     115-14 Hospital 15 Baseline visit Demographics   age          -
#> 151     115-24 Hospital 15 Baseline visit Demographics   age          -
#> 152      116-6  Hospital 4 Baseline visit Demographics   age          -
#> 153     116-12  Hospital 4 Baseline visit Demographics   age          -
#> 154     116-14  Hospital 4 Baseline visit Demographics   age          -
#> 155      117-6  Hospital 2 Baseline visit Demographics   age          -
#> 156     117-12  Hospital 2 Baseline visit Demographics   age          -
#> 157     117-25  Hospital 2 Baseline visit Demographics   age          -
#> 158     117-26  Hospital 2 Baseline visit Demographics   age          -
#> 159      118-7 Hospital 22 Baseline visit Demographics   age          -
#> 160     118-10 Hospital 22 Baseline visit Demographics   age          -
#> 161     118-12 Hospital 22 Baseline visit Demographics   age          -
#> 162      119-6  Hospital 7 Baseline visit Demographics   age          -
#> 163      119-7  Hospital 7 Baseline visit Demographics   age          -
#> 164     119-14  Hospital 7 Baseline visit Demographics   age          -
#> 165      120-6 Hospital 12 Baseline visit Demographics   age          -
#> 166     120-10 Hospital 12 Baseline visit Demographics   age          -
#> 167     120-13 Hospital 12 Baseline visit Demographics   age          -
#> 168     120-15 Hospital 12 Baseline visit Demographics   age          -
#> 169     120-26 Hospital 12 Baseline visit Demographics   age          -
#> 170      121-7  Hospital 3 Baseline visit Demographics   age          -
#> 171     121-10  Hospital 3 Baseline visit Demographics   age          -
#> 172      122-8 Hospital 16 Baseline visit Demographics   age          -
#> 173     122-12 Hospital 16 Baseline visit Demographics   age          -
#> 174     122-13 Hospital 16 Baseline visit Demographics   age          -
#> 175     122-18 Hospital 16 Baseline visit Demographics   age          -
#> 176     122-19 Hospital 16 Baseline visit Demographics   age          -
#> 177     122-23 Hospital 16 Baseline visit Demographics   age          -
#> 178     122-30 Hospital 16 Baseline visit Demographics   age          -
#> 179     122-33 Hospital 16 Baseline visit Demographics   age          -
#> 180     122-35 Hospital 16 Baseline visit Demographics   age          -
#> 181     123-10  Hospital 9 Baseline visit Demographics   age          -
#> 182      124-7 Hospital 17 Baseline visit Demographics   age          -
#> 183      124-9 Hospital 17 Baseline visit Demographics   age          -
#> 184      125-6 Hospital 23 Baseline visit Demographics   age          -
#> 185     125-10 Hospital 23 Baseline visit Demographics   age          -
#>     Description                                    Query      Code
#> 1           Age The value is 56 and it should not be >20   100-6-1
#> 2           Age The value is 60 and it should not be >20  100-13-1
#> 3           Age The value is 68 and it should not be >20  100-16-1
#> 4           Age The value is 75 and it should not be >20  100-31-1
#> 5           Age The value is 65 and it should not be >20  100-34-1
#> 6           Age The value is 67 and it should not be >20  100-36-1
#> 7           Age The value is 65 and it should not be >20  100-52-1
#> 8           Age The value is 72 and it should not be >20  100-58-1
#> 9           Age The value is 76 and it should not be >20  100-65-1
#> 10          Age The value is 75 and it should not be >20  100-66-1
#> 11          Age The value is 76 and it should not be >20  100-72-1
#> 12          Age The value is 76 and it should not be >20  100-82-1
#> 13          Age The value is 61 and it should not be >20   101-8-1
#> 14          Age The value is 36 and it should not be >20  101-13-1
#> 15          Age The value is 78 and it should not be >20  101-16-1
#> 16          Age The value is 74 and it should not be >20  101-19-1
#> 17          Age The value is 68 and it should not be >20  101-22-1
#> 18          Age The value is 83 and it should not be >20  101-36-1
#> 19          Age The value is 64 and it should not be >20  101-45-1
#> 20          Age The value is 58 and it should not be >20  101-48-1
#> 21          Age The value is 85 and it should not be >20  101-59-1
#> 22          Age The value is 74 and it should not be >20  101-70-1
#> 23          Age The value is 70 and it should not be >20  101-74-1
#> 24          Age The value is 71 and it should not be >20  101-75-1
#> 25          Age The value is 62 and it should not be >20  101-78-1
#> 26          Age The value is 77 and it should not be >20  101-83-1
#> 27          Age The value is 62 and it should not be >20  101-85-1
#> 28          Age The value is 64 and it should not be >20  101-86-1
#> 29          Age The value is 71 and it should not be >20  101-88-1
#> 30          Age The value is 72 and it should not be >20  101-93-1
#> 31          Age The value is 64 and it should not be >20  101-96-1
#> 32          Age The value is 42 and it should not be >20  101-98-1
#> 33          Age The value is 53 and it should not be >20   102-6-1
#> 34          Age The value is 86 and it should not be >20  102-10-1
#> 35          Age The value is 63 and it should not be >20  102-13-1
#> 36          Age The value is 58 and it should not be >20  102-15-1
#> 37          Age The value is 68 and it should not be >20  102-23-1
#> 38          Age The value is 71 and it should not be >20  102-27-1
#> 39          Age The value is 75 and it should not be >20  102-30-1
#> 40          Age The value is 63 and it should not be >20  102-39-1
#> 41          Age The value is 52 and it should not be >20  102-48-1
#> 42          Age The value is 83 and it should not be >20  102-60-1
#> 43          Age The value is 70 and it should not be >20  102-64-1
#> 44          Age The value is 74 and it should not be >20  102-73-1
#> 45          Age The value is 57 and it should not be >20  102-74-1
#> 46          Age The value is 76 and it should not be >20  102-75-1
#> 47          Age The value is 80 and it should not be >20  102-83-1
#> 48          Age The value is 78 and it should not be >20  102-84-1
#> 49          Age The value is 55 and it should not be >20 102-103-1
#> 50          Age The value is 55 and it should not be >20 102-104-1
#> 51          Age The value is 83 and it should not be >20 102-127-1
#> 52          Age The value is 61 and it should not be >20 102-134-1
#> 53          Age The value is 61 and it should not be >20   103-6-1
#> 54          Age The value is 86 and it should not be >20   103-9-1
#> 55          Age The value is 81 and it should not be >20  103-10-1
#> 56          Age The value is 52 and it should not be >20  103-14-1
#> 57          Age The value is 75 and it should not be >20  103-16-1
#> 58          Age The value is 21 and it should not be >20  103-17-1
#> 59          Age The value is 79 and it should not be >20  103-19-1
#> 60          Age The value is 62 and it should not be >20   104-8-1
#> 61          Age The value is 64 and it should not be >20  104-12-1
#> 62          Age The value is 47 and it should not be >20  104-13-1
#> 63          Age The value is 71 and it should not be >20   105-8-1
#> 64          Age The value is 84 and it should not be >20   105-9-1
#> 65          Age The value is 89 and it should not be >20  105-13-1
#> 66          Age The value is 83 and it should not be >20  105-19-1
#> 67          Age The value is 78 and it should not be >20  105-21-1
#> 68          Age The value is 68 and it should not be >20  105-22-1
#> 69          Age The value is 82 and it should not be >20  105-30-1
#> 70          Age The value is 76 and it should not be >20  105-46-1
#> 71          Age The value is 71 and it should not be >20  105-52-1
#> 72          Age The value is 73 and it should not be >20  105-58-1
#> 73          Age The value is 50 and it should not be >20  105-78-1
#> 74          Age The value is 80 and it should not be >20  105-85-1
#> 75          Age The value is 66 and it should not be >20   106-8-1
#> 76          Age The value is 53 and it should not be >20   106-9-1
#> 77          Age The value is 67 and it should not be >20  106-15-1
#> 78          Age The value is 67 and it should not be >20  106-18-1
#> 79          Age The value is 84 and it should not be >20  106-22-1
#> 80          Age The value is 66 and it should not be >20  106-27-1
#> 81          Age The value is 66 and it should not be >20  106-29-1
#> 82          Age The value is 68 and it should not be >20  106-32-1
#> 83          Age The value is 73 and it should not be >20  106-33-1
#> 84          Age The value is 87 and it should not be >20  107-11-1
#> 85          Age The value is 85 and it should not be >20   108-8-1
#> 86          Age The value is 57 and it should not be >20  108-13-1
#> 87          Age The value is 54 and it should not be >20  108-16-1
#> 88          Age The value is 59 and it should not be >20  108-17-1
#> 89          Age The value is 57 and it should not be >20  108-19-1
#> 90          Age The value is 72 and it should not be >20  108-23-1
#> 91          Age The value is 65 and it should not be >20  108-28-1
#> 92          Age The value is 43 and it should not be >20  108-29-1
#> 93          Age The value is 49 and it should not be >20  108-30-1
#> 94          Age The value is 73 and it should not be >20  108-35-1
#> 95          Age The value is 69 and it should not be >20  108-44-1
#> 96          Age The value is 58 and it should not be >20  108-45-1
#> 97          Age The value is 64 and it should not be >20  108-56-1
#> 98          Age The value is 70 and it should not be >20  108-63-1
#> 99          Age The value is 47 and it should not be >20   109-8-1
#> 100         Age The value is 77 and it should not be >20  109-10-1
#> 101         Age The value is 83 and it should not be >20  109-11-1
#> 102         Age The value is 68 and it should not be >20  109-19-1
#> 103         Age The value is 54 and it should not be >20  109-21-1
#> 104         Age The value is 91 and it should not be >20  109-22-1
#> 105         Age The value is 78 and it should not be >20  109-32-1
#> 106         Age The value is 69 and it should not be >20  109-33-1
#> 107         Age The value is 63 and it should not be >20  109-34-1
#> 108         Age The value is 45 and it should not be >20  109-35-1
#> 109         Age The value is 69 and it should not be >20  109-37-1
#> 110         Age The value is 85 and it should not be >20  109-38-1
#> 111         Age The value is 40 and it should not be >20   110-6-1
#> 112         Age The value is 90 and it should not be >20  110-19-1
#> 113         Age The value is 57 and it should not be >20   111-7-1
#> 114         Age The value is 62 and it should not be >20   111-8-1
#> 115         Age The value is 58 and it should not be >20  111-10-1
#> 116         Age The value is 56 and it should not be >20   112-6-1
#> 117         Age The value is 89 and it should not be >20   112-7-1
#> 118         Age The value is 76 and it should not be >20  112-10-1
#> 119         Age The value is 51 and it should not be >20  112-11-1
#> 120         Age The value is 74 and it should not be >20  112-12-1
#> 121         Age The value is 75 and it should not be >20  112-14-1
#> 122         Age The value is 55 and it should not be >20  112-15-1
#> 123         Age The value is 71 and it should not be >20  112-18-1
#> 124         Age The value is 49 and it should not be >20  112-19-1
#> 125         Age The value is 63 and it should not be >20  112-20-1
#> 126         Age The value is 59 and it should not be >20  112-21-1
#> 127         Age The value is 62 and it should not be >20   113-9-1
#> 128         Age The value is 51 and it should not be >20  113-10-1
#> 129         Age The value is 61 and it should not be >20  113-11-1
#> 130         Age The value is 35 and it should not be >20  113-12-1
#> 131         Age The value is 49 and it should not be >20  113-13-1
#> 132         Age The value is 41 and it should not be >20  113-16-1
#> 133         Age The value is 74 and it should not be >20  113-21-1
#> 134         Age The value is 73 and it should not be >20  113-23-1
#> 135         Age The value is 88 and it should not be >20  113-24-1
#> 136         Age The value is 47 and it should not be >20  113-27-1
#> 137         Age The value is 63 and it should not be >20  113-31-1
#> 138         Age The value is 80 and it should not be >20  113-32-1
#> 139         Age The value is 89 and it should not be >20  113-36-1
#> 140         Age The value is 79 and it should not be >20  113-37-1
#> 141         Age The value is 69 and it should not be >20  113-38-1
#> 142         Age The value is 78 and it should not be >20  113-40-1
#> 143         Age The value is 78 and it should not be >20  113-41-1
#> 144         Age The value is 80 and it should not be >20  113-44-1
#> 145         Age The value is 78 and it should not be >20  114-13-1
#> 146         Age The value is 66 and it should not be >20  114-17-1
#> 147         Age The value is 79 and it should not be >20  114-19-1
#> 148         Age The value is 68 and it should not be >20   115-8-1
#> 149         Age The value is 31 and it should not be >20  115-13-1
#> 150         Age The value is 65 and it should not be >20  115-14-1
#> 151         Age The value is 62 and it should not be >20  115-24-1
#> 152         Age The value is 36 and it should not be >20   116-6-1
#> 153         Age The value is 51 and it should not be >20  116-12-1
#> 154         Age The value is 68 and it should not be >20  116-14-1
#> 155         Age The value is 55 and it should not be >20   117-6-1
#> 156         Age The value is 72 and it should not be >20  117-12-1
#> 157         Age The value is 67 and it should not be >20  117-25-1
#> 158         Age The value is 69 and it should not be >20  117-26-1
#> 159         Age The value is 71 and it should not be >20   118-7-1
#> 160         Age The value is 27 and it should not be >20  118-10-1
#> 161         Age The value is 79 and it should not be >20  118-12-1
#> 162         Age The value is 56 and it should not be >20   119-6-1
#> 163         Age The value is 68 and it should not be >20   119-7-1
#> 164         Age The value is 81 and it should not be >20  119-14-1
#> 165         Age The value is 68 and it should not be >20   120-6-1
#> 166         Age The value is 79 and it should not be >20  120-10-1
#> 167         Age The value is 76 and it should not be >20  120-13-1
#> 168         Age The value is 56 and it should not be >20  120-15-1
#> 169         Age The value is 64 and it should not be >20  120-26-1
#> 170         Age The value is 70 and it should not be >20   121-7-1
#> 171         Age The value is 54 and it should not be >20  121-10-1
#> 172         Age The value is 70 and it should not be >20   122-8-1
#> 173         Age The value is 46 and it should not be >20  122-12-1
#> 174         Age The value is 57 and it should not be >20  122-13-1
#> 175         Age The value is 47 and it should not be >20  122-18-1
#> 176         Age The value is 70 and it should not be >20  122-19-1
#> 177         Age The value is 26 and it should not be >20  122-23-1
#> 178         Age The value is 53 and it should not be >20  122-30-1
#> 179         Age The value is 77 and it should not be >20  122-33-1
#> 180         Age The value is 80 and it should not be >20  122-35-1
#> 181         Age The value is 64 and it should not be >20  123-10-1
#> 182         Age The value is 72 and it should not be >20   124-7-1
#> 183         Age The value is 77 and it should not be >20   124-9-1
#> 184         Age The value is 68 and it should not be >20   125-6-1
#> 185         Age The value is 48 and it should not be >20  125-10-1
#> 
#> $results
#> <table class="table table-striped table-condensed" style="width: auto !important; margin-left: auto; margin-right: auto;">
#> <caption>Report of queries</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Variables </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Description </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Event </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Query </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Total </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:center;"> age </td>
#>    <td style="text-align:center;"> Age </td>
#>    <td style="text-align:center;"> Baseline visit </td>
#>    <td style="text-align:center;"> The value should not be &gt;20 </td>
#>    <td style="text-align:center;"> 185 </td>
#>   </tr>
#> </tbody>
#> </table>

# Using the filter argument
example <- rd_query(covican,
                    variables = "potassium",
                    expression = "%in%NA",
                    event = "baseline_visit_arm_1",
                    filter = "available_analytics=='1'")
#> Warning: At least one of the variables that have been checked for missing values present a branching logic. 
#> Check the results element of the output(...$results) for details.
example
#> $queries
#>    Identifier         DAG          Event          Instrument     Field
#> 1       102-6 Hospital 24 Baseline visit Laboratory findings potassium
#> 2      102-13 Hospital 24 Baseline visit Laboratory findings potassium
#> 3      102-23 Hospital 24 Baseline visit Laboratory findings potassium
#> 4      102-27 Hospital 24 Baseline visit Laboratory findings potassium
#> 5      102-30 Hospital 24 Baseline visit Laboratory findings potassium
#> 6      102-39 Hospital 24 Baseline visit Laboratory findings potassium
#> 7      102-48 Hospital 24 Baseline visit Laboratory findings potassium
#> 8      102-60 Hospital 24 Baseline visit Laboratory findings potassium
#> 9      102-64 Hospital 24 Baseline visit Laboratory findings potassium
#> 10     102-73 Hospital 24 Baseline visit Laboratory findings potassium
#> 11     102-75 Hospital 24 Baseline visit Laboratory findings potassium
#> 12     102-84 Hospital 24 Baseline visit Laboratory findings potassium
#> 13    102-103 Hospital 24 Baseline visit Laboratory findings potassium
#> 14    102-104 Hospital 24 Baseline visit Laboratory findings potassium
#> 15    102-113 Hospital 24 Baseline visit Laboratory findings potassium
#> 16    102-127 Hospital 24 Baseline visit Laboratory findings potassium
#> 17      111-7 Hospital 14 Baseline visit Laboratory findings potassium
#> 18      111-8 Hospital 14 Baseline visit Laboratory findings potassium
#> 19     113-11  Hospital 8 Baseline visit Laboratory findings potassium
#> 20      120-6 Hospital 12 Baseline visit Laboratory findings potassium
#> 21      125-6 Hospital 23 Baseline visit Laboratory findings potassium
#>    Repetition Description                                        Query
#> 1           -   Potassium The value is NA and it should not be missing
#> 2           -   Potassium The value is NA and it should not be missing
#> 3           -   Potassium The value is NA and it should not be missing
#> 4           -   Potassium The value is NA and it should not be missing
#> 5           -   Potassium The value is NA and it should not be missing
#> 6           -   Potassium The value is NA and it should not be missing
#> 7           -   Potassium The value is NA and it should not be missing
#> 8           -   Potassium The value is NA and it should not be missing
#> 9           -   Potassium The value is NA and it should not be missing
#> 10          -   Potassium The value is NA and it should not be missing
#> 11          -   Potassium The value is NA and it should not be missing
#> 12          -   Potassium The value is NA and it should not be missing
#> 13          -   Potassium The value is NA and it should not be missing
#> 14          -   Potassium The value is NA and it should not be missing
#> 15          -   Potassium The value is NA and it should not be missing
#> 16          -   Potassium The value is NA and it should not be missing
#> 17          -   Potassium The value is NA and it should not be missing
#> 18          -   Potassium The value is NA and it should not be missing
#> 19          -   Potassium The value is NA and it should not be missing
#> 20          -   Potassium The value is NA and it should not be missing
#> 21          -   Potassium The value is NA and it should not be missing
#>         Code
#> 1    102-6-1
#> 2   102-13-1
#> 3   102-23-1
#> 4   102-27-1
#> 5   102-30-1
#> 6   102-39-1
#> 7   102-48-1
#> 8   102-60-1
#> 9   102-64-1
#> 10  102-73-1
#> 11  102-75-1
#> 12  102-84-1
#> 13 102-103-1
#> 14 102-104-1
#> 15 102-113-1
#> 16 102-127-1
#> 17   111-7-1
#> 18   111-8-1
#> 19  113-11-1
#> 20   120-6-1
#> 21   125-6-1
#> 
#> $results
#> <table class="table table-striped table-condensed" style="width: auto !important; margin-left: auto; margin-right: auto;">
#> <caption>Report of queries</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Variables </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Description </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Event </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Query </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Total </th>
#>    <th style="text-align:center;border-bottom: 1px solid grey"> Branching logic </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:center;"> potassium </td>
#>    <td style="text-align:center;"> Potassium </td>
#>    <td style="text-align:center;"> Baseline visit </td>
#>    <td style="text-align:center;"> The value should not be missing </td>
#>    <td style="text-align:center;"> 21 </td>
#>    <td style="text-align:center;"> [available_analytics]='1' </td>
#>   </tr>
#> </tbody>
#> </table>